+++ src/view/components/elements/Overlay.vue
@@ -1,19 +1,37 @@
 <template>
-  <div class='ui overlay' transition='fade' @click.self='shadeClicked'>
-    <div class='ui panel' :style="{ width: panelWidth, height: panelHeight }">
+  <section class='ui overlay' transition='fade' @click.self='close'>
+    <section class='ui panel' :style="{ width: panelWidth, height: panelHeight }">
       <slot></slot>
-    </div>
-  </div>
+      <button v-if="!noClose" class='ui subtle hover-darken close-button' @click='close'>
+        <i class='fa fa-times'></i>
+      </button>
+    </section>
+  </section>
 </template>
+
+<style lang="stylus" scoped>
+.panel
+  position: relative
+
+.close-button
+  max-width: 100vw
+  max-height: 100vh
+  width: 2em
+  height: @width
+  background: none
+  position: absolute
+  top: 0
+  right: 0
+</style>
 
 <script>
 import {PopOverlay} from '../../lib/events'
 
 export default {
   props: {
-    closeOnShadeClicked: {
+    noClose: {
       type: Boolean,
-      default: true
+      default: false
     },
 
     panelWidth: {
@@ -23,19 +41,15 @@ export default {
 
     panelHeight: {
       type: String,
-      default: 'auto'
+      default: 'max-content'
     }
   },
 
   methods: {
-    shadeClicked () {
-      if (this.closeOnShadeClicked) {
-        this.close()
-      }
-    },
-
     close () {
-      this.$dispatch(PopOverlay)
+      if (!this.noClose) {
+        this.$dispatch(PopOverlay)
+      }
     }
   }
 }